MGDAPI-5809 Renamed monitoringcommon to obo
