+++ pkg/products/obo/alertmanagerConfig.go
@@ -1,4 +1,4 @@
-package monitoringcommon
+package obo
 
 import (
 	"context"
