+++ pkg/products/obo/alertmanagerConfig_test.go
@@ -1,4 +1,4 @@
-package monitoringcommon
+package obo
 
 import (
 	"bytes"
