+++ pkg/products/obo/prometheusRules.go
@@ -1,4 +1,4 @@
-package monitoringcommon
+package obo
 
 import (
 	"fmt"
