+++ pkg/products/obo/templateHelper.go
@@ -1,4 +1,4 @@
-package monitoringcommon
+package obo
 
 import (
 	"bytes"
