+++ pkg/products/observability/reconciler.go
@@ -6,7 +6,7 @@ import (
 	integreatlyv1alpha1 "github.com/integr8ly/integreatly-operator/apis/v1alpha1"
 	"github.com/integr8ly/integreatly-operator/pkg/addon"
 	"github.com/integr8ly/integreatly-operator/pkg/config"
-	"github.com/integr8ly/integreatly-operator/pkg/products/monitoringcommon"
+	"github.com/integr8ly/integreatly-operator/pkg/products/obo"
 	"github.com/integr8ly/integreatly-operator/pkg/resources"
 	"github.com/integr8ly/integreatly-operator/pkg/resources/events"
 	l "github.com/integr8ly/integreatly-operator/pkg/resources/logger"
@@ -146,7 +146,7 @@ func (r *Reconciler) deleteObservabilityCR(ctx context.Context, serverClient k8s
 	}
 	if isHiveManaged {
 		// proceed after the dms secret is deleted by the deadmanssnitch-operator to prevent alert false positive
-		dmsSecret, err := monitoringcommon.GetDMSSecret(ctx, serverClient, *inst)
+		dmsSecret, err := obo.GetDMSSecret(ctx, serverClient, *inst)
 		if err != nil && !k8serr.IsNotFound(err) {
 			return integreatlyv1alpha1.PhaseFailed, fmt.Errorf("unexpected error retrieving dead man's snitch secret: %w", err)
 		}
